move main.py to src and rename it __init__.py
