+++ smart_library_start.bat
@@ -1,4 +1,4 @@
 call  C:\Intel\computer_vision_sdk_2018.5.456\bin\setupvars.bat
-cd src/GUI/
-python main.py
+cd src
+python __init__.py
 pause
