+++ src/infrastructure/Database/Readers.csv
@@ -1,3 +1,2 @@
 user_id,book_id,borrow_date,return_date
-2,2,0911,-1
-3,3,2311,-1
+
