+++ src/infrastructure/Database/Users/UserRole.csv
@@ -1,4 +1,2 @@
 user_id,role_id
-1,1
-2,1
-3,2
+
