+++ src/infrastructure/Database/Users/Users.csv
@@ -1,4 +1,2 @@
 user_id,phone,first_name,last_name,middle_name
-1,+7831,A1,B1,C1
-2,+7831,Medved,Medvedev,Medvedovich
-3,+7831,A3,B3,C3
+1,+0000000000,Ivan,Vikhrev,Borisovich
